Renamed lp.module.scss to about.module.scss
