+++ src/styles/pages/_index.scss
@@ -1,3 +1,3 @@
-@forward 'lp.module';
+@forward 'about.module';
 @forward 'mood.module';
 @forward 'customize.module';
