--- src/styles/pages/_lp.scss
@@ -1,4 +1,0 @@
-@use '../abstracts' as a;
-@use '../mixin/' as m;
-
-.lp {}
